requirements.txt
Morfessor==2.0.6
numpy==1.18.4
oseti==0.2
polyglot==16.7.4
pycld2==0.41
PyICU==2.4.3
pymlask==0.3.2
responder==2.0.5
